refactor: change names files and change method renderNewPage
