+++ src/core/components/footer/footer.ts
