+++ src/core/components/header/header.template.ts
@@ -1,0 +1,1 @@
+// const HEADERTEMLATE = ``;
