+++ src/core/components/header/header.ts
@@ -1,0 +1,5 @@
+// import stringToElement from '../../../utils/htmlToElement';
+
+// const headerElement: HTMLTemplateElement = stringToElement(HEADERTEMLATE);
+
+// export default headerElement;
