--- src/core/components/header/index.scss
@@ -1,7 +1,0 @@
-.header__nav {
-  width: fit-content;
-  margin: auto;
-  a + a {
-    margin-left: 40px;
-  }
-}
