+++ src/index.ts
@@ -1,8 +1,6 @@
 import './index.scss';
-import App from './pages/app';
+import App from './pages/app/app';
 
 const app = new App();
 
 app.run();
-
-console.log(`I'm working...`);
